requirements.txt
playwright==1.54.0
python-dotenv==1.0.1
supabase==2.7.4
pandas==2.2.2
openpyxl==3.1.5
pyarrow==16.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
# Gmail API dependencies for draft generator
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
